Complete the initial TripVault class and test - moved method from the Trips class to TripVault
